+++ _stub_w3m.js
@@ -1,0 +1,3 @@
+export const createWeb3Modal = () => {};
+export const useWeb3Modal = () => ({ open: () => {} });
+export const Web3Modal = () => null;